+++ src/session/termcode.ml
@@ -25,6 +25,12 @@ let arg_extra_expl_prefix =
    Arg.String (fun s -> expl_prefixes := s :: !expl_prefixes),
    "<s> register s as an additional prefix for VC explanations")
 
+let opt_extra_expl_prefix =
+  let open Getopt in
+  KLong "extra-expl-prefix",
+  Hnd1 (AString, fun s -> expl_prefixes := s :: !expl_prefixes),
+  "<expl> register <expl> as an additional prefix\nfor VC explanations"
+
 let collect_expls attr =
   Ident.Sattr.fold
     (fun attr acc ->