+++ src/tools/why3shell.ml
@@ -381,14 +381,15 @@ let files = Queue.create ()
 
 let quiet = ref false
 
-let set_quiet () = quiet := true
-
 let spec =
-  ["--quiet", Arg.Unit set_quiet, "Remove all printing to stdout"]
+  let open Getopt in
+  [KLong "quiet", Hnd0 (fun () -> quiet := true),
+   " remove all printing to stdout"]
 
 (* --help *)
-let usage_str = sprintf
-  "Usage: %s [options] [ <file.xml> | <f1.why> <f2.mlw> ...]"
+let usage_str = Format.sprintf
+  "Usage: %s [options] [ <file.xml> | <f1.why> <f2.mlw> ...]\n\
+   Launch a command-line interface for Why3.\n"
   (Filename.basename Sys.argv.(0))
 
 (* Parse files *)